Always explicitly specify methods for routes
This revealled a bug. We were doing 302 redirects previously, which
meant that after PUTting a new project the redirect to reload the page
came as another PUT. After explicitly specifying GET only for the
index page that PUT on the index failed (previously, when unspecified,
any method was allowed). So now we do a 303 redirect instead which
switches to GET and is intended for exactly this case.